Crate optimization

Re-exports

pub use self::compile_time_assertions::CompileTimeAssertions;

pub use self::compile_time_assertions::Phase;

Modules

compile_time_assertions
Injects the class 'Assert' into the mini java runtime, which contains functions that are evaluated at compile-time.

Structs

CommonSubExpr

Optimization

OutcomeCollector

RemoveCriticalEdges

Enums

Flag
These are options that can be attached to any optimization

Kind
All available optimizations

Level
These are predefined sequences of optimizations (in clang and gcc these are called -O1, -O3, -Os and so forth).

Outcome

Traits

Interprocedural
An optimization that optimizes the whole program by examining all function graphs at once.

Local
An optimization that only works on a single graph and therefore does not optimize across function call boundaries.